Backend_FastAPI/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Authentication & Security (no compilation needed)
PyJWT==2.10.1
bcrypt==4.0.1

# Database
motor==3.3.2
pymongo==4.11.2

# Audio Processing (keep from Flask)
audioop-lts==0.2.1
audioread==3.0.1
librosa==0.11.0
soundfile==0.13.1
soxr==0.5.0.post1

# ML & Data Processing (keep from Flask)
torch==2.6.0
numpy==2.1.3
opencv-python==4.11.0.86
scikit-learn==1.6.1
scipy==1.15.2

# Async File Operations
aiofiles==23.2.0

# Environment & Config
python-dotenv==1.0.1
pydantic==2.5.0

# Support Libraries (keep from Flask)
numba==0.61.0
llvmlite==0.44.0
joblib==1.4.2
decorator==5.2.1
packaging==24.2
requests==2.32.3

# Production Server
gunicorn==23.0.0